+++ src/main/java/ch/fhnw/deardevbackend/mapper/WorkKindInsightMapper.java
@@ -15,17 +15,17 @@ public interface WorkKindInsightMapper {
     @Mapping(source = "userCount", target = "userCount")
     @Mapping(source = "teamAverage", target = "teamAverage")
     @Mapping(source = "teamCount", target = "teamCount")
-    WorkKindInsightDTO toDTO(Integer workKindId, String workKindName, Double userAverage, Long userCount, Double teamAverage, Long teamCount);
+    WorkKindInsightDTO toDTO(Integer workKindId, String workKindName, Integer userAverage, Integer userCount, Integer teamAverage, Integer teamCount);
 
     @Mapping(source = "workKindId", target = "workKindId")
     @Mapping(source = "workKindName", target = "workKindName")
     @Mapping(source = "userAverage", target = "userAverage")
     @Mapping(source = "userCount", target = "userCount")
-    WorkKindInsightDTO toUserDTO(Integer workKindId, String workKindName, Double userAverage, Long userCount);
+    WorkKindInsightDTO toUserDTO(Integer workKindId, String workKindName, Integer userAverage, Integer userCount);
 
     @Mapping(source = "workKindId", target = "workKindId")
     @Mapping(source = "workKindName", target = "workKindName")
     @Mapping(source = "teamAverage", target = "teamAverage")
     @Mapping(source = "teamCount", target = "teamCount")
-    WorkKindInsightDTO toTeamDTO(Integer workKindId, String workKindName, Double teamAverage, Long teamCount);
+    WorkKindInsightDTO toTeamDTO(Integer workKindId, String workKindName, Integer teamAverage, Integer teamCount);
 }
